feat: support grayscale images

+++ circle_evolution/fitness.py
@@ -26,8 +26,7 @@ class Fitness:
         Args:
             target (np.ndarray): target image array.
         """
-        self.target = target
-        self.target = self.target.astype(np.float32)
+        self.target = target.astype(np.float32)
 
     def score(self, phenotype):
         """Score a Specie.
